Fix the type conversion warnings.

+++ astreamer/audio_queue.h
@@ -62,13 +62,13 @@ private:
     AudioQueueBufferRef m_audioQueueBuffer[AQ_BUFFERS];              // audio queue buffers	
     AudioStreamPacketDescription m_packetDescs[AQ_MAX_PACKET_DESCS]; // packet descriptions for enqueuing audio
     
-    size_t m_fillBufferIndex;                                        // the index of the audioQueueBuffer that is being filled
-    size_t m_bytesFilled;                                            // how many bytes have been filled
-    size_t m_packetsFilled;                                          // how many packets have been filled
-    size_t m_buffersUsed;                                            // how many buffers are used
+    UInt32 m_fillBufferIndex;                                        // the index of the audioQueueBuffer that is being filled
+    UInt32 m_bytesFilled;                                            // how many bytes have been filled
+    UInt32 m_packetsFilled;                                          // how many packets have been filled
+    UInt32 m_buffersUsed;                                            // how many buffers are used
     
-    size_t m_processedPacketsSizeTotal;                              // global packet statistics: total size
-    size_t m_processedPacketsCount;                                  // global packet statistics: count
+    UInt32 m_processedPacketsSizeTotal;                              // global packet statistics: total size
+    UInt32 m_processedPacketsCount;                                  // global packet statistics: count
     
     bool m_audioQueueStarted;                                        // flag to indicate that the queue has been started
     bool m_bufferInUse[AQ_BUFFERS];                                  // flags to indicate that a buffer is still in use
